+++ flogin/jsonrpc/__init__.py
@@ -2,3 +2,4 @@ from .client import *
 from .errors import *
 from .responses import *
 from .results import *
+from .glyph import *